chore: auto-generate version details (if possible)
Instead of having a static version details, automated the generation of
the version detail strings to be presented to the user. This
auto-generation works only if the source code has been 'git cloned.'

Technically the source code can be released as a zip file (directly)
from github. In this case, there will be no 'git' data available and
hence no version information. To mitigate this, I was forced to hardcode
the current version details inside 'version.mk' file. This needs to be
modified everytime we need to tag! I am not sure how this is going to
work! I would prefer if everybody could git clone & build the final
binary!